betty again

+++ 0x00-hello_world/6-size.c
@@ -4,7 +4,7 @@ int main(void)
 {
 	int a;
 	long int b;
-	long int c;
+	long long int c;
 	char d;
 	float f;
